NEW SKIN PAGE
as the title says, i made a better skin page, i made a cool css check it if you want, i also added some utilitys. You can now get the skin from another player by getting their name and theorically use an url (i dodn't tested it)., Their is still a bug in the pages that switchs the informations with an other account which is annoying im trying to stop that in the next update. Also i used div and a grid generator to make a good ui. I hope you'll like it